First milestone: join procedure working!
TODO: 1. handle devNonce during the join process
2. handle mapping of the datarate for downlink RX1Window
3. *IMPORTANT* Think about whether the server should use
UDP or TCP to talk to the gateway. The current implementation
uses TCP but in a broken/naive way -- does not check for
data completeness, as it assumes the data arrives at one piece.
Should either switch to UDP, or changes to protocol to add
a length field to each datagram.